Apply more code review fixes
