+++ dali/kernels/signal/fft/fft_cpu_impl_utils.h
@@ -18,7 +18,6 @@
 #include <utility>
 #include <vector>
 #include "dali/core/util.h"
-#include "dali/kernels/signal/signal_kernel_utils.h"
 
 namespace dali {
 namespace kernels {
@@ -64,39 +63,6 @@ struct MagnitudeSpectrumCalculator {
     }
   }
 };
-
-/**
- * @brief iterator through all the 1-dimensional slices on a given axis
- */
-template <typename OutputType, typename InputType, typename Functor>
-void ForAxis(OutputType *out_ptr,
-             const InputType *in_ptr,
-             const int64_t *out_shape,
-             const int64_t *out_strides,
-             const int64_t *in_shape,
-             const int64_t *in_strides,
-             int axis,
-             int ndim,
-             Functor &&func,
-             int current_dim = 0) {
-  if (current_dim == ndim) {
-    func(out_ptr, in_ptr, out_shape[axis], out_strides[axis], in_shape[axis], in_strides[axis]);
-    return;
-  }
-
-  if (axis == current_dim) {
-    ForAxis(out_ptr, in_ptr, out_shape, out_strides, in_shape, in_strides,
-            axis, ndim, std::forward<Functor>(func), current_dim+1);
-  } else {
-    for (int i = 0; i < in_shape[current_dim]; i++) {
-      ForAxis(out_ptr + i * out_strides[current_dim],
-              in_ptr + i * in_strides[current_dim],
-              out_shape, out_strides,
-              in_shape, in_strides,
-              axis, ndim, std::forward<Functor>(func), current_dim+1);
-    }
-  }
-}
 
 }  // namespace impl
 }  // namespace fft
